+++ src/main/java/pl/plajer/villagedefense/creatures/upgrades/Upgrade.java
@@ -19,10 +19,11 @@
 package pl.plajer.villagedefense.creatures.upgrades;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
-import org.bukkit.entity.Entity;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -46,16 +47,19 @@ public class Upgrade {
   private List<String> description;
   private String metadataAccessor;
   private String configAccessor;
+  private Map<Integer, Double> tieredValues = new HashMap<>();
 
-  public Upgrade(String id, int slot, EntityType entityType, int maxTier, String name, List<String> description, String metadataAccessor, String configAccessor) {
+  public Upgrade(String id) {
     this.id = id;
-    this.slot = slot;
+    this.configAccessor = "Entity-Upgrades." + id + "-Tiers";
+  }
+
+  public void setEntityType(EntityType entityType) {
     this.entityType = entityType;
-    this.maxTier = maxTier;
-    this.name = name;
-    this.description = description;
-    this.metadataAccessor = metadataAccessor;
-    this.configAccessor = configAccessor;
+  }
+
+  public void setTierValue(int tier, double value) {
+    tieredValues.put(tier, value);
   }
 
   public String getId() {
@@ -66,6 +70,10 @@ public class Upgrade {
     return slot;
   }
 
+  public void setSlot(int slot) {
+    this.slot = slot;
+  }
+
   public EntityType getApplicableFor() {
     return entityType;
   }
@@ -74,69 +82,56 @@ public class Upgrade {
     return maxTier;
   }
 
+  public void setMaxTier(int maxTier) {
+    this.maxTier = maxTier;
+  }
+
   public String getName() {
     return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
   }
 
   public List<String> getDescription() {
     return description;
   }
 
+  public void setDescription(List<String> description) {
+    this.description = description;
+  }
+
   public String getMetadataAccessor() {
     return metadataAccessor;
+  }
+
+  public void setMetadataAccessor(String metadataAccessor) {
+    this.metadataAccessor = metadataAccessor;
+  }
+
+  public double getValueForTier(int tier) {
+    return tieredValues.getOrDefault(tier, tieredValues.get(maxTier));
   }
 
   public int getCost(int tier) {
     return plugin.getConfig().getInt(configAccessor + "." + tier);
   }
 
-  //todo bit weird, maybe recode
-  public ItemStack asItemStack(Entity en, int currentTier, Upgrade upgrade) {
+  public ItemStack asItemStack(int currentTier) {
     String name = this.name;
     List<String> description = new ArrayList<>(this.description);
 
-    String from;
-    String to;
-    int tier = plugin.getEntityUpgradeMenu().getTier(en, upgrade);
-    switch (upgrade.getId()) {
-      case "DAMAGE":
-        from = String.valueOf(2.0 + (tier * 3));
-        to = String.valueOf(2.0 + ((tier + 1) * 3));
-        break;
-      case "HEALTH":
-        from = String.valueOf(100.0 + (100.0 * ((double) tier / 2.0)));
-        to = String.valueOf(100.0 + (100.0 * ((double) (tier + 1) / 2.0)));
-        break;
-      case "SPEED":
-        from = String.valueOf(0.25 + (0.25 * ((double) tier / 5.0)));
-        to = String.valueOf(0.25 + (0.25 * ((double) (tier + 1) / 5.0)));
-        break;
-      case "SWARM_AWARENESS":
-        from = String.valueOf(tier * 0.2);
-        to = String.valueOf((tier + 1) * 0.2);
-        break;
-      case "FINAL_DEFENSE":
-        from = String.valueOf(tier * 5);
-        to = String.valueOf((tier + 1) * 5);
-        break;
-      default:
-        from = "";
-        to = "";
-        break;
-    }
-    if (tier == upgrade.getMaxTier()) {
-      to = from;
-    }
-
-    final String finalTo = to;
+    double valCurrent = tieredValues.get(currentTier);
+    double valNext = tieredValues.getOrDefault(currentTier + 1, tieredValues.get(currentTier));
 
     description = description.stream().map((lore) -> lore = plugin.getChatManager().colorRawMessage(lore)
         .replace("%cost%", String.valueOf(getCost(currentTier + 1)))
         .replace("%tier%", String.valueOf(currentTier + 1))
-        .replace("%from%", from)
-        .replace("%to%", finalTo)).collect(Collectors.toList());
+        .replace("%from%", String.valueOf(valCurrent))
+        .replace("%to%", String.valueOf(valNext))).collect(Collectors.toList());
     return new ItemBuilder(XMaterial.BLACK_STAINED_GLASS_PANE.parseItem())
-        .name(plugin.getChatManager().colorRawMessage(name))
+        .name(name)
         .lore(description).build();
   }